fixed double

+++ app.py
@@ -24,8 +24,8 @@ class users(Base):
 class locations(Base):
    __tablename__ = "locations" 
    user_id = Column(Integer, primary_key = True)
-   latitude = Column(Double)
-   longitude = Column(Double) 
+   latitude = Column(Float)
+   longitude = Column(Float) 
 
 
 class friends(Base):
